fix(feature-engineer): cast Quantity to float64 before per-unit division
np.where's result dtype for revenue_per_unit and profit_per_unit
depended on the installed pandas/numpy version when dividing by the
nullable Int64 Quantity column: pandas 3.0.2/numpy 2.4.4 produced a
float64 array, but pandas 2.1.4/numpy 1.26.4 (pinned inside the
Airflow runtime image) produced an object-dtype array, causing
.round(2) to raise TypeError during a live Airflow DAG run.

Cast Quantity to float64 before both divisions so the result dtype
is stable regardless of installed pandas/numpy version. Adds a
regression test covering a nullable Int64 Quantity column with a
zero-quantity row.

+++ src/transform/feature_engineer.py
@@ -145,18 +145,23 @@ def add_financial_features(df: pd.DataFrame) -> pd.DataFrame:
     # revenue_per_unit: average revenue generated per unit sold.
     # Formula: Sales / Quantity
     # np.where guards against division by zero when Quantity == 0.
+    # Quantity is cast to float64 before dividing so the division always runs
+    # on a native numpy float array — nullable "Int64" Quantity otherwise makes
+    # np.where's result dtype depend on the installed pandas/numpy version,
+    # which breaks .round(2) on older pinned versions (e.g. Airflow's image).
     df["revenue_per_unit"] = np.where(
         df["Quantity"] != 0,  # Condition: at least one unit sold
-        df["Sales"] / df["Quantity"],  # True: divide revenue by quantity
+        df["Sales"] / df["Quantity"].astype("float64"),  # True: divide revenue by quantity
         0.0,  # False: default to 0 for safety
     ).round(2)
 
     # profit_per_unit: average net profit (or loss) per unit sold.
     # Formula: Profit / Quantity
     # A negative value means this product is sold at a loss per unit.
+    # Quantity is cast to float64 for the same dtype-stability reason as above.
     df["profit_per_unit"] = np.where(
         df["Quantity"] != 0,  # Condition: at least one unit sold
-        df["Profit"] / df["Quantity"],  # True: divide profit by quantity
+        df["Profit"] / df["Quantity"].astype("float64"),  # True: divide profit by quantity
         0.0,  # False: default to 0 for safety
     ).round(2)
